the arduous process continues
