+++ src/App.vue
@@ -4,12 +4,14 @@ import MenuBar from './components/Menus/MenuBar.vue';
 import Footer from './components/Footer.vue';
 import { defineComponent } from "vue";
 import { useUserStore } from './stores/user';
+import { getActivePinia, MutationType } from "pinia";
 
 export default defineComponent({
   setup() {
       const user = useUserStore();
-      //user.$subscribe()
-    return { user }
+      const pinia = getActivePinia()
+    
+    return { user, pinia }
   },
   components: {
     MenuBar,
@@ -17,19 +19,15 @@ export default defineComponent({
     Footer
   },
   watch: {
-    /*
-    "character": {
+    "pinia.state.value.character": {
+      // sync the user's open character with the app's openCharacter
       deep: true,
-      handler: function (after) {
-        localStorage.setItem("character", JSON.stringify(after));
-      },
+      handler: function () {
+        this.user.$patch((state) => {
+          state.openCharacter = this.pinia.state.value.character
+        })
+      }
     },
-     "savedCharacters": {
-      deep: true,
-      handler: function (after) {
-        localStorage.setItem("savedCharacters", JSON.stringify(after));
-      },
-    }, */ 
   },  
   mounted() {
   }
@@ -37,7 +35,7 @@ export default defineComponent({
 </script>
 
 <template>
-  <div class="max-w-md mb-4 pb-4 mx-auto bg-white rounded-xl shadow-md md:max-w-2xl">
+  <div class="max-w-md mb-4 pb-4 mx-auto text-center bg-white rounded-xl shadow-md md:max-w-2xl">
     <MenuBar :user="user" />
     <RouterView />
   </div>
@@ -47,12 +45,5 @@ export default defineComponent({
 <style>
 #app {
   font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
 }
 </style>
-
-function mutation(mutation: any, state: any) {
-  throw new Error("Function not implemented.");
-}
