+++ src/components/Menus/MenuActual.vue
@@ -1,3 +1,46 @@
+<template>   
+    <div class="z-50 text-left relative" v-clickOutside="hideMenu">
+
+
+        <MenuBarButton 
+        
+            :label="menu.label" 
+        
+        
+            @button-event="toggleShowMenu()" 
+        />    
+
+        <transition
+            enter-active-class="duration-100 ease-out"
+            enter-from-class="transform opacity-0 -translate-y-6"
+            enter-to-class="opacity-100"
+            leave-active-class="duration-75 ease-in"
+            leave-from-class="opacity-100"
+            leave-to-class="transform opacity-0 -translate-y-6"
+        >
+        <ul 
+            class="flex flex-col absolute top-11 left-2
+            border-[1px] p-2 bg-slate-50 rounded-sm min-w-fit"
+
+
+            v-if="showMenuValue"
+
+        >
+            <li 
+            
+                v-for="(item, index) in menu.items" 
+                
+                
+                :key="index">
+                <MenuActualButton :item="item" @hide-menu="hideMenu" />
+            </li>
+        </ul>
+        
+        </transition>
+
+    </div>
+</template>
+
 <script lang="ts">
 import { defineComponent } from "vue";
 import MenuBarButton from './MenuBarButton.vue'
@@ -15,10 +58,7 @@ export default defineComponent({
         MenuActualButton
     },
     setup() {
-        const MenuOptions: Array<String> = ['new character','save character']
-        return {
-            MenuOptions
-        }
+
     },      
     methods: {
         toggleShowMenu() { this.showMenuValue = !this.showMenuValue; },
@@ -28,32 +68,3 @@ export default defineComponent({
     },
 })
 </script>
-
-<template>   
-    <div class="z-50 text-left relative" v-clickOutside="hideMenu">
-
-
-        <MenuBarButton :label="menu.label" @button-event="toggleShowMenu()" />    
-
-        <transition
-            enter-active-class="duration-100 ease-out"
-            enter-from-class="transform opacity-0 -translate-y-6"
-            enter-to-class="opacity-100"
-            leave-active-class="duration-75 ease-in"
-            leave-from-class="opacity-100"
-            leave-to-class="transform opacity-0 -translate-y-6"
-        >
-        <ul 
-            class="flex flex-col absolute top-11 left-2
-            border-[1px] p-2 bg-slate-50 rounded-sm min-w-fit"
-            v-if="showMenuValue"
-        >
-            <li v-for="(item, index) in menu.items" :key="index">
-                <MenuActualButton :item="item" @hide-menu="hideMenu" />
-            </li>
-        </ul>
-        
-        </transition>
-
-    </div>
-</template>
