+++ src/stores/user.ts
@@ -17,6 +17,10 @@ export interface User {
   undoHistory: [],
 }
 
+interface SavedCharacters {
+  [index: number]: Character
+}
+
 
 interface UserLocalStorage {
   useStorage(s: string,user: User): Ref<User>,
@@ -26,124 +30,19 @@ interface UserLocalStorage {
 export const useUserStore = defineStore({
   id: "user",
   state: () => ({
-    userLocalStorage: useStorage(
-      'my-toh-skill-builder', {
-      id: String,
-      username: String,
-      openCharacter: Object,
-      savedCharacters: [],
-      prefs: {
-        localStorageSize: 0,
-      },
-    }
-    ),
+    id: String,
+    username: String,
+    openCharacter: Object,
+    savedCharacters: [],
+    prefs: {
+      localStorageSize: 0,
+    },
   }) as StateTree,
   getters: {
 
   },
   actions: {
-    getUser() {
-      /** remove old skills & replace with new character.skills */
-      //if (this.character.skills = JSON.parse(localStorage.getItem("skills"))) { localStorage.removeItem("skills") } 
-      /** if exists, sync browser with localStorage */
-      //if (this.character.totalPoints = JSON.parse(localStorage.getItem("total"))) { localStorage.removeItem("total") }
-      //if this.isVariableUnset(this.character.totalPoints))
-      /* below: original uncommented toh-skill-builder code for mounted() {}:
-      this.character = JSON.parse(localStorage.getItem("character")) || this.character
-      this.savedCharacters = JSON.parse(localStorage.getItem("savedCharacters")) || this.savedCharacters
-      */
-    },
-
-    deleteAllMyStuff() {
-
-      // done? maybe force a reload      
-      // remove all proof of this app's existence from your device, for a few milliseconds.
-      // old data is replaced with a blank slate.
-      // (emergencies only, cannot be undone)
-      // localStorage.removeItem('my-toh-skill-builder'); 
-      this.$reset()
-
-    },         
-    migrateOldLocalStorageCharacters() {
-      // what's in localStorage?
-      // console.log('localStorage: ' + JSON.stringify(localStorage))    
-      // which permutation does it have?
-      // code
-      // copy those permutations to the new model  
-      // code
-      //trying to get here:
-      /*localStorage.removeItem('skills');      
-      localStorage.removeItem('total');      
-      localStorage.removeItem('character');      
-      localStorage.removeItem('savedCharacters');
-      localStorage.removeItem('size');*/
-    },
-    setLocalStorageSize() {
-      console.log('localStorage will be set here in the future')
-      // for use when it's time to implement an undo... I waste dtoo much time on it today
-      /*
-      console.log(JSON.stringify(this.$state.userLocalStorage.prefs))
-      console.log(/*JSON.stringify/this.userLocalStorage.prefs.localStorageSize)
-      const userLocalStorage: UserLocalStorage = this.userLocalStorage
-      //this.deleteAllMyStuff();
-      // if localStoage exists, and user's cache is set to zero (the default),
-      /*
-      if (localStorage && ((this.$state.userLocalStorage.prefs.localStorageSize === 0) || (this.$state.userLocalStorage.prefs.localStorageSize === undefined))) {
-        var i = 0;
-        try {
-            // test up to 10 MB 
-            for (i = 250; i <= 10000; i += 250) {
-                localStorage.setItem('test', new Array((i * 1024) + 1).join('a'));
-            }
-        } catch (e) {
-            // delete test object
-            localStorage.removeItem('test');
-            // set user's localStorage pref, 
-            this.$state.prefs.localStorageSize = i - 250;    
-        }        
-      }
-      else {
-        // no localstorage or cache size alrady set
-      }
-      */
-    },
-    createCharacter(character: Character) { 
-      //const characterStore = useCharacter()
-      //const savedCharacters: array = this.userLocalStorage.savedCharacters)
-      //this.userLocalStorage.savedCharacters.push(character); 
-      console.log("creating to be implemented") 
-    },    
-    saveCharacter() {
-      console.log('saving to be implemented')
-      // if character is unnamed, give name "Unnamed (x)"
-      /*if (this.isVariableUnset(this.character.name)) { 
-        var unnamedArray = this.savedCharacters.filter(function (character) { return character.name.includes("Unnamed ")})      
-        var unnamedIndex = unnamedArray.length
-        this.character.name = "Unnamed " + unnamedIndex
-      }
-      console.log(this.character.name + " saving")
-        if (this.isVariableUnset(this.character.id)) {
-          this.character.id = Math.random().toString(36).slice(2);
-          console.log("id generated " + this.character.id)
-          this.createCharacter(this.character);
-        } 
-        else {
-          console.log("id is " + this.character.id + " and tested valid")
-          var savedIndex = this.savedCharacters.findIndex(obj => { return obj.id === this.character.id } ) 
-          if (savedIndex === -1) { 
-            console.log("...but character did not exist in savedCharacters. \n create ")
-            this.createCharacter(this.character) 
-          }
-          else { 
-            console.log("update")
-            this.updateCharacter(savedIndex)
-          }
-        }
-        console.log("saved")
-      },
-      */      
-    },
-    openCharacter() {},
+    saveCharacter(character: any) { this.state.savedCharacters.push(character) }
   },
 });
